refactor(notification): extract useScrollThumb, fix ghost thumb, throttle scroll updates
Address AI review findings on the custom scrollbar: the thumb stayed
visible with stale coordinates when the scrollable list unmounted out
from under it (e.g. status flips to loading/empty while the popover is
still open), scroll events drove a React state update on every fire
instead of once per frame, and the hook was missing cleanup test
coverage. Also move useScrollThumb out of NotificationBell.tsx into its
own file (src/hooks/useScrollThumb.ts) since it's a generic DOM
concern unrelated to notification business logic.
