+++ src/components/layout/Header/NotificationBell.tsx
@@ -15,6 +15,7 @@ import {
   type NotificationItem,
   useNotificationBell,
 } from '@/hooks/useNotificationBell';
+import { useScrollThumb } from '@/hooks/useScrollThumb';
 import { formatRelativeTime } from '@/lib/dateUtils';
 import { cn } from '@/lib/utils';
 
@@ -49,170 +50,6 @@ export type NotificationBellProps = {
    */
   onMarkAllRead?: (ids: string[]) => void | Promise<void>;
 };
-
-const SCROLL_THUMB_MIN_HEIGHT = 24;
-const SCROLL_THUMB_HIDE_DELAY = 1000;
-
-type ScrollThumbState = {
-  /** Whether the content actually overflows (whether to render at all). */
-  visible: boolean;
-  /** Whether the thumb should currently be opaque (hover or recent scroll). */
-  active: boolean;
-  top: number;
-  height: number;
-};
-
-type ScrollThumbHandlers = {
-  trackRefCallback: React.RefCallback<HTMLDivElement>;
-  scrollRefCallback: React.RefCallback<HTMLDivElement>;
-  onMouseEnter: () => void;
-  onMouseLeave: () => void;
-};
-
-/**
- * Tracks scroll position on the scrollable list and derives a thumb
- * top/height mapped onto the height of a larger "track" element (the
- * whole card, header included), so a custom (non-native) scrollbar
- * indicator can be rendered spanning from the very top of the card.
- * Native `::-webkit-scrollbar-*` overrides are unreliable on some
- * current Windows Chromium builds (Fluent scrollbars can ignore
- * `::-webkit-scrollbar-button { display: none }`), so this sidesteps
- * native scrollbar rendering entirely.
- *
- * The thumb fades out after a period of no hover/scroll activity, and
- * reappears on hover or scroll, matching overlay-scrollbar conventions.
- *
- * Uses callback refs rather than plain `useRef` + `useEffect`: content
- * renders inside a Radix `Popover` portal, which can mount on a
- * different timing than the parent's own effect cycle. A callback ref
- * is guaranteed by React to fire exactly when its DOM node
- * mounts/unmounts, regardless of that.
- */
-function useScrollThumb(): [ScrollThumbHandlers, ScrollThumbState] {
-  const [geometry, setGeometry] = React.useState<
-    Omit<ScrollThumbState, 'active'>
-  >({ visible: false, top: 0, height: 0 });
-  const [active, setActive] = React.useState(false);
-
-  const trackElRef = React.useRef<HTMLDivElement | null>(null);
-  const scrollElRef = React.useRef<HTMLDivElement | null>(null);
-  const cleanupRef = React.useRef<(() => void) | null>(null);
-  const hideTimeoutRef = React.useRef<ReturnType<typeof setTimeout> | null>(
-    null
-  );
-  const isHoveredRef = React.useRef(false);
-  const wasOverflowingRef = React.useRef(false);
-
-  const scheduleHide = React.useCallback(() => {
-    if (hideTimeoutRef.current) clearTimeout(hideTimeoutRef.current);
-    hideTimeoutRef.current = setTimeout(() => {
-      if (!isHoveredRef.current) setActive(false);
-    }, SCROLL_THUMB_HIDE_DELAY);
-  }, []);
-
-  const onMouseEnter = React.useCallback(() => {
-    isHoveredRef.current = true;
-    if (hideTimeoutRef.current) clearTimeout(hideTimeoutRef.current);
-    setActive(true);
-  }, []);
-
-  const onMouseLeave = React.useCallback(() => {
-    isHoveredRef.current = false;
-    if (hideTimeoutRef.current) clearTimeout(hideTimeoutRef.current);
-    setActive(false);
-  }, []);
-
-  const recompute = React.useCallback(() => {
-    const trackEl = trackElRef.current;
-    const scrollEl = scrollElRef.current;
-    if (!trackEl || !scrollEl) return;
-
-    const { scrollTop, scrollHeight, clientHeight } = scrollEl;
-    const isOverflowing = scrollHeight > clientHeight;
-
-    if (!isOverflowing) {
-      wasOverflowingRef.current = false;
-      setGeometry({ visible: false, top: 0, height: 0 });
-      return;
-    }
-
-    const trackHeight = trackEl.clientHeight;
-    const height = Math.max(
-      (clientHeight / scrollHeight) * trackHeight,
-      SCROLL_THUMB_MIN_HEIGHT
-    );
-    const maxTop = trackHeight - height;
-    const scrollableDistance = scrollHeight - clientHeight;
-    const top =
-      scrollableDistance > 0 ? (scrollTop / scrollableDistance) * maxTop : 0;
-    setGeometry({ visible: true, top, height });
-
-    // Briefly reveal as an affordance the first time content becomes
-    // scrollable (e.g. right when the popover opens).
-    if (!wasOverflowingRef.current) {
-      wasOverflowingRef.current = true;
-      setActive(true);
-      scheduleHide();
-    }
-  }, [scheduleHide]);
-
-  const attachListeners = React.useCallback(() => {
-    const trackEl = trackElRef.current;
-    const scrollEl = scrollElRef.current;
-    cleanupRef.current?.();
-    cleanupRef.current = null;
-    if (!trackEl || !scrollEl) return;
-
-    const handleScroll = () => {
-      recompute();
-      setActive(true);
-      scheduleHide();
-    };
-
-    recompute();
-    scrollEl.addEventListener('scroll', handleScroll, { passive: true });
-
-    // jsdom (unit tests) doesn't implement ResizeObserver.
-    const resizeObserver =
-      typeof ResizeObserver !== 'undefined'
-        ? new ResizeObserver(recompute)
-        : null;
-    resizeObserver?.observe(trackEl);
-    resizeObserver?.observe(scrollEl);
-
-    cleanupRef.current = () => {
-      scrollEl.removeEventListener('scroll', handleScroll);
-      resizeObserver?.disconnect();
-    };
-  }, [recompute, scheduleHide]);
-
-  const trackRefCallback = React.useCallback(
-    (el: HTMLDivElement | null) => {
-      trackElRef.current = el;
-      attachListeners();
-    },
-    [attachListeners]
-  );
-
-  const scrollRefCallback = React.useCallback(
-    (el: HTMLDivElement | null) => {
-      scrollElRef.current = el;
-      attachListeners();
-    },
-    [attachListeners]
-  );
-
-  React.useEffect(() => {
-    return () => {
-      if (hideTimeoutRef.current) clearTimeout(hideTimeoutRef.current);
-    };
-  }, []);
-
-  return [
-    { trackRefCallback, scrollRefCallback, onMouseEnter, onMouseLeave },
-    { ...geometry, active },
-  ];
-}
 
 /**
  * Returns content templates (title, body) for notification items.
